+++ src/components/radio-group/radio-group-interface.ts
@@ -1,3 +1,5 @@
 export interface RadioGroupChangeEventDetail {
     value: any | any[] | undefined | null;
 }
+
+export type RadioGroupMode = 'radio' | 'option';