+++ stories/checkbox/checkbox.stories.ts
@@ -14,8 +14,12 @@ export default {
     },
 } as Meta;
 
-const Template: Story<Partial<{ checked: boolean, indeterminate: boolean, name: string, label: string, labelAlignment: CheckboxLabelAlignment }>> = (args) => html`
-    <wcs-checkbox ?checked=${args.checked} ?indeterminate=${args.indeterminate} name=${args.name} label-alignment=${args.labelAlignment}>
+const Template: Story<Partial<{ checked: boolean, indeterminate: boolean, name: string, label: string, labelAlignment: CheckboxLabelAlignment, disabled: boolean }>> = (args) => html`
+    <wcs-checkbox ?checked=${args.checked}
+                  ?indeterminate=${args.indeterminate}
+                  name=${args.name}
+                  label-alignment=${args.labelAlignment}
+                  ?disabled=${args.disabled}>
         ${args.label}
     </wcs-checkbox>
 `;
@@ -26,5 +30,16 @@ Default.args = {
     indeterminate: false,
     name: 'checkbox-id',
     label: 'Une checkbox',
-    labelAlignment: 'center'
+    labelAlignment: 'center',
+    disabled: false
+};
+
+export const Disabled = Template.bind({});
+Disabled.args = {
+    checked: false,
+    indeterminate: false,
+    name: 'checkbox-id',
+    label: 'Une checkbox',
+    labelAlignment: 'center',
+    disabled: true
 };